BugTracker bugId=420
Scale the X axis so that it always shows a min of zero

Altered WaterChartControl.cs, edited the properties so that the minimum value of the top axis was 0 and the maximum value was automatic.
While I was there I also made the left axis maximum value automatic instead of 1180mm.