Bug 411
Fixed this by in FileUI OnSave just added a test to see if the file exists. 

Add a ToolTip to the Search button to explain what it does. 

After fixing this, I noticed that when in the development toolbox you can browse to a different .ini file and it will display it, but once you have dragged it into the simulation then you can no longer browse to a different .ini file. The "\" disappeared from the path, and nothing was displayed. 

This is caused by if you have not saved the simulation yet, then Controller.ApsimData.FileName is null, so when you do a replace, you just replace all the "\" with "" instead. Fixed this with an if statement.